Add more primitive types, including integer, floating-point, character and enable index support in strings.

+++ interpreter/BWHazel.TopsyTurvy.Tests/Parser/TopsyTurvyParserArrayTests.cs
@@ -58,7 +58,7 @@ public class TopsyTurvyParserArrayTests
     /// <param name="expectedType">The expected element <see cref="LiteralType"/>.</param>
     [Theory]
     [InlineData("PEER",   LiteralType.Integer)]
-    [InlineData("FATHOM", LiteralType.Float)]
+    [InlineData("FATHOM", LiteralType.Double)]
     [InlineData("YARN",   LiteralType.String)]
     [InlineData("DECREE", LiteralType.Boolean)]
     [InlineData("NAUGHT", LiteralType.Null)]
